requirements_chatbot.txt
openai-whisper>=20240314
pyttsx3>=2.90
pyaudio>=0.2.13
chromadb>=0.3.21
indic-nlp-library>=0.71
langdetect>=1.0.9
pandas>=1.3.0
numpy>=1.21.0
groq>=0.4.2
python-dotenv>=0.19.0
